backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-cloud-vision>=3.10.2",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "pytesseract>=0.3.13",
    "httpx>=0.28.1",
]
